+++ app/Http/Controllers/PageSpeedController.php
@@ -45,22 +45,4 @@ class PageSpeedController extends Controller
         curl_exec($ch);
         return new JsonResponse(curl_getinfo($ch));
     }
-
-    public function verifyEmail(Request $request)
-    {
-        if (! $request->hasValidSignature()) {
-            abort(401);
-        }
-        $code = $request->get('code');
-        $evc = EmailVerificationCode::where('code', $code)->first();
-        if ($evc instanceof EmailVerificationCode) {
-            $evc->date_email_confirmed = Carbon::now();
-            $page = tap($evc->page()->get())->date_email_confirmed = Carbon::now();
-            $page->save();
-            $evc->save();
-            echo "SUCCESS!";
-        } else {
-            abort(402);
-        }
-    }
 }
