+++ routes/web.php
@@ -15,10 +15,10 @@ Route::get('/', function () {
     return view('dashboard');
 })->name('home');
 
-Route::get('/pagespeed', 'PageSpeedController@index');
-Route::post('/page/validate', 'PageSpeedController@validateUrl');
+Route::resource('/pagespeed', 'PageSpeedController');
+Route::post('/pagespeed/validate', 'PageSpeedController@validateUrl');
 Route::get('/pagespeed/faq', function () {
     return view('pagespeed.faq');
 })->name('pagespeed.faq');
 
-Route::get('/pagespeed/verify', 'PageSpeedController@verifyEmail')->name('pagespeed.verifyEmail');
+Route::get('/verify-email', 'EmailVerificationCodeController@verifyEmail')->name('pagespeed.verifyEmail');
